NavMeshMovementBasics
Imported navmesh into the project and updated states code to work with it

Idle:
Every so often, pick a spot biased towards the player and set a destination to it

Chase;
Set a destination towards the player.  If you're in the sensor radius, do this every frame, if you're outside, do this every time the player exceeds a certain distance away form the destination

MoveToPoint
Set a destination to a starting point once and head straight there

Note:
Wander is a little buggy, as the bias is returning nonsensical points after referencing the correct corner of the path.