General movement: Created better detection for if player is about to hit the ground out of a half pipe & fixed most slope clipping issues
Added even more raycasts that I call "extensions". They come out of the same point as the slope orientation raycasts, but they go forward in their respective cardinal directions to tell the movement code when the player is about to smack onto the ground, so corrections can be made early.