bulletInformation and States firing bullets
BulletInformation is a script attached to an empty object on the enemy. This allows us to define multiple kinds of bullets.

Added enemy turret state and gave bullet firing logic to both it and the chase state

TODO: put spawning code on the bulletInfo class so we can extend it with different kinds of bullets