Start: Animation Implementation on ESDG states
Enemy states now tie to an animation on entering, which they will use to play.

Added AnimationEventHandler which will bet put on the object with the animator. Animation Events will be caught by this handler and outputted elsewhere as a UnityEvent

TODO: Add Event Functionality to EventHandler
TODO: Remove bullet firing logic from turret and chase, and assign it to playing the fireball animation.